refactor: move showcasecolormode
