+++ src/types/metadata.d.ts
@@ -1,4 +1,4 @@
-import { ShowcaseColorMode } from 'lib/characterPreview/ShowcaseCustomizationSidebar'
+import { ShowcaseColorMode } from 'lib/constants/constants'
 import { SortOptionProperties } from 'lib/optimization/sortOptions'
 import { PresetDefinition } from 'lib/tabs/tabOptimizer/optimizerForm/components/RecommendedPresetsButton'
